远上寒山石径斜，白云生处有人家
暂时不使用预先画出地面的方法
场景不自己创建纹理，都由Engine创建
加快回合制的退场渐变

+++ src/BattleScene.h
@@ -60,7 +60,7 @@ protected:
     int show_number_y_ = 0;
     int effect_id_ = -1;
     int effect_frame_ = 0;
-    uint8_t dead_alpha_ = 255;
+    int dead_alpha_ = 255;
     static const int animation_delay_ = 2;
 
     // 动画，或者需要多写一坨if else等
